SpiralIterator Test, for center out of bounds
The test was created because initializing a SpiralIterator with a center that is
outside of the map was leading to a segmentation fault when accessing the iterator.

The test creates a SpiralIterator with a center that is outside of the map with
a radius chosen, so that the iterator will reach the map.
Then the created iterator is used to get the corresponding position on the map
and it is verified that the returned position is actually inside the map.